requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
pydantic-settings
structlog==23.2.0

# Asynchronous Operations
aiofiles==23.2.1
aiohttp==3.9.1

# --- AI & Machine Learning Core ---
# Install PyTorch separately using the official command (see guide below)
# torch==2.6.0
# torchaudio==2.6.0
llama-cpp-python==0.2.20
sentence-transformers==2.7.0
langchain-core==0.1.41
langchain-community==0.0.32
langchain-text-splitters==0.0.1
scikit-learn==1.5.2
numpy>=1.26,<2.0
yake==0.4.8
rich==13.7.0 


# Natural Language Processing (NLP)
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9
textblob==0.17.1
textdistance==4.6.3
joblib==1.3.2

# Vector Database & Search
faiss-cpu==1.7.4

# --- Document Processing ---
PyMuPDF==1.22.5
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10 # Requires external Tesseract installation
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
markdown==3.5.1
python-magic-bin==0.4.14 # Windows compatible version of python-magic
rarfile==4.1
py7zr==0.20.6
pypdf2==3.0.1  
ebooklib==0.18 
pandas==2.2.2 
# --- Voice & Audio Processing ---
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
noisereduce==2.0.1
scipy==1.11.4
webrtcvad-wheels==2.0.11

# --- Caching System ---
redis==5.0.1
aioredis==2.0.1
diskcache==5.6.3
lz4==4.3.2
orjson==3.9.10
msgpack==1.0.7

# --- Utilities & Monitoring ---
psutil==5.9.6
prometheus-client==0.19.0
networkx==3.2.1
textstat